split DocumentEditor/component-blocks/api into shared/ui
