+++ packages/fields-document/src/DocumentEditor/component-blocks/component-block-render.tsx
@@ -4,7 +4,7 @@ import { jsx } from '@keystone-ui/core'
 import React, { useContext } from 'react'
 import { useMemo, type ReactElement } from 'react'
 import { type Element } from 'slate'
-import { type ComponentBlock } from './api'
+import { type ComponentBlock } from './api-shared'
 import { createGetPreviewProps, getKeysForArrayValue } from './preview-props'
 import { type ReadonlyPropPath } from './utils'
