+++ packages/fields-document/src/DocumentEditor/insert-menu.tsx
@@ -12,7 +12,7 @@ import { ReactEditor } from 'slate-react'
 import { matchSorter } from 'match-sorter'
 import scrollIntoView from 'scroll-into-view-if-needed'
 import { ComponentBlockContext, insertComponentBlock } from './component-blocks'
-import { type ComponentBlock } from './component-blocks/api'
+import { type ComponentBlock } from './component-blocks/api-shared'
 import { InlineDialog, ToolbarButton } from './primitives'
 import { type Relationships, useDocumentFieldRelationships } from './relationship'
 import { type ToolbarState, useToolbarState } from './toolbar-state'
