+++ packages/fields-document/src/structure-graphql-output.tsx
@@ -1,6 +1,6 @@
 import { graphql } from '@keystone-6/core'
 import { type FieldData } from '@keystone-6/core/types'
-import { type ComponentSchemaForGraphQL } from './DocumentEditor/component-blocks/api'
+import { type ComponentSchemaForGraphQL } from './DocumentEditor/component-blocks/api-shared'
 import { assertNever } from './DocumentEditor/component-blocks/utils'
 
 function wrapGraphQLFieldInResolver<InputSource, OutputSource> (
