+++ packages/fields-document/src/structure-validation.ts
@@ -1,6 +1,6 @@
 import * as t from 'io-ts'
 import excess from 'io-ts-excess'
-import { type RelationshipData } from './DocumentEditor/component-blocks/api'
+import { type RelationshipData } from './DocumentEditor/component-blocks/api-shared'
 import { type Mark } from './DocumentEditor/utils'
 import { isValidURL } from './DocumentEditor/isValidURL'
 // note that this validation isn't about ensuring that a document has nodes in the right positions and things
